Make sensor tolerate 'no such process' errors during tracing (#342)
For ptrace(), it's normal not to find a traced process from time to
time. For instance, it could happen when the tracee is KILL-ed. If
such an event happens while the master process is still awaiting the
sensor's startup, it would cause the `docker-slim build` command to
fail with -124 exit code leaving the sensor in the (semi-)initialized
state (and not killing the container).

This commit prevents the sensor from sending any 'no such process'
events as errors back to the master. Instead such events will be logged
on the sensor side.