Healthcheck fixes (#332)
* Add unit tests for deserialiseHealtheckInstruction

* Do not add default values for HEALTHCHECK to the reconstructed one

* Correct HEALTHCHECK parsing

deserialiseHealtheckInstruction is expecting to get the full history line and
not just the parameters

* Fix the parsing of retries in HEALTHCHECK

the --retries value in the history is formatted via %q, which means that we have
to treat the cases of the number being printed as a hex value, a unicode value,
a character or a C escape sequence separately

* Fix formatting of the --retries parameter for HEALTHCHECK

* update vendored dependencies

* Fix reconstruction of the CMD-SHELL HEALTHCHECK

the config.Test when HEALTHCHECK is using the shell form is expected to be a two
element array: the first entry is `CMD-SHELL` and the second is the full
command. We were erronously using the split value here instead of joined one.