Standalone Sensor Mode - Allow overriding app_name and app_args for instrumented images
The instrumented image must have the `commands.json` file baked in
(based on the image's ENTRYPOINT and/or CMD). However, it's also
possible to override the CMD by passing arguments to the `docker run` command.
Since these args can't be known at runtime, sensor should be able to
detect them and replace the `app_name` and `app_args` fields in the
monitor start command it read from the baked in `commands.json` file.