feat: Standalone Sensor Mode (#383)
Sensor got two new flags (the exact names to be refined thouhg):

-mode=controlled|standalone
-command=<JSON-encoded MonitorStart command>

The default mode is 'controlled' and sensor's behavior in that mode
(supposedly) stayed unchanged.

In the 'standalone' mode, sensor becomes independent from the driver app
(docker-slim). Sensor terminates as long as the target app is terminated
and the artifacts have been persisted.

Additionally, the signal forwarding (from the sensor's process to the
target app) is implemented, including sending SIGKILL to the target app
after a grace period. Beware, though, that the signal forwarding works
only in the standalone mode.